# Conflicts:
#	app/src/main/java/com/towich/cosmicintrigue/ui/adapters/VoteAdapter.kt
#	app/src/main/java/com/towich/cosmicintrigue/ui/fragment/MapFragment.kt

+++ app/src/main/java/com/towich/cosmicintrigue/ui/adapters/WaitAdapter.kt
@@ -10,7 +10,7 @@ import com.towich.cosmicintrigue.databinding.WroomViewBinding
 import com.towich.cosmicintrigue.data.model.Player
 import com.towich.cosmicintrigue.data.model.ReadyPlayer
 
-class WaitAdapter(val userId : Int) :RecyclerView.Adapter<WaitAdapter.WaitViewHolder>(){
+class WaitAdapter(val userId : Long) :RecyclerView.Adapter<WaitAdapter.WaitViewHolder>(){
     class WaitViewHolder(val binding: WroomViewBinding):RecyclerView.ViewHolder(binding.root)
     var players : List<ReadyPlayer> = ArrayList()
     fun setReady(player: List<ReadyPlayer>){
@@ -30,7 +30,7 @@ class WaitAdapter(val userId : Int) :RecyclerView.Adapter<WaitAdapter.WaitViewHo
             var p = players.get(position)
             this.loginText.text = p.login
             this.readyText.text = if(p.ready) "Готов" else "Не готов"
-            if(p.id == userId)
+            if(p.id == userId.toInt())
                 this.back.setBackgroundColor(Color.parseColor("#F0F0F0"))
             this.readyText.setTextColor(if(p.ready) Color.parseColor("#FF0000") else Color.parseColor("#00FF00"))
         }
